+++ mayan/apps/document_indexing/forms.py
@@ -4,29 +4,23 @@ from django import forms
 from django.utils.encoding import force_text
 from django.utils.translation import ugettext_lazy as _
 
-from mayan.apps.acls.models import AccessControlList
 from mayan.apps.common.classes import ModelProperty
+from mayan.apps.common.forms import FilteredSelectionForm
 from mayan.apps.documents.models import Document
 
 from .models import Index, IndexTemplateNode
 from .permissions import permission_document_indexing_rebuild
 
 
-class IndexListForm(forms.Form):
-    indexes = forms.ModelMultipleChoiceField(
-        help_text=_('Indexes to be queued for rebuilding.'),
-        label=_('Indexes'), queryset=Index.objects.none(),
-        required=False, widget=forms.widgets.CheckboxSelectMultiple()
-    )
-
-    def __init__(self, *args, **kwargs):
-        user = kwargs.pop('user')
-        super(IndexListForm, self).__init__(*args, **kwargs)
-        queryset = AccessControlList.objects.filter_by_access(
-            permission=permission_document_indexing_rebuild, user=user,
-            queryset=Index.objects.filter(enabled=True)
-        )
-        self.fields['indexes'].queryset = queryset
+class IndexTemplateFilteredForm(FilteredSelectionForm):
+    class Meta:
+        allow_multiple = True
+        field_name = 'index_templates'
+        help_text = _('Index templates to be queued for rebuilding.')
+        label = _('Index templates')
+        queryset = Index.objects.filter(enabled=True)
+        permission = permission_document_indexing_rebuild
+        widget_attributes = {'class': 'select2'}
 
 
 class IndexTemplateNodeForm(forms.ModelForm):